#Fix dataBase

+++ BaseLibrary/Entities/Employee.cs
@@ -8,7 +8,7 @@ namespace BaseLibrary.Entities
 {
     public class Employee: BaseEntity
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public string? CivilId { get; set; }
         public string? FileNumber { get; set; }
@@ -21,15 +21,15 @@ namespace BaseLibrary.Entities
 
         //RelationShip: Many to One
         public GeneralDepartment? GeneralDepartment { get; set; }
-        public long GeneralDepartmentId { get; set; }
+        public int GeneralDepartmentId { get; set; }
 
         public Department? Department { get; set; }
-        public long DepartmentId { get; set; }
+        public int DepartmentId { get; set; }
 
         public Branch? Branch { get; set; }
-        public long BranchId { get; set; }
+        public int BranchId { get; set; }
 
         public Town? Town { get; set; }
-        public long TownId { get; set; }
+        public int TownId { get; set; }
     }
 }
